// apps/electron/src/renderer/src/app/types.ts
export interface PluginInstance {
  id: string
  getPath: () => Promise<string | null>
  isActive: () => Promise<boolean>
  onStateChange: (callback: (enabled: boolean) => void) => () => void
  command: (commandId: string, ...args: any[]) => Promise<any>
  invoke: (channel: string, ...args: any[]) => Promise<any>
  on: (channel: string, callback: (...args: any[]) => void) => void
  off: (channel: string, callback: (...args: any[]) => void) => void
  enable: () => Promise<void>
  disable: () => Promise<void>
}

export interface PluginAPI {
  get: (pluginId: string) => PluginInstance
  list: () => Promise<any[]>
  install: () => Promise<any>
  loadDev: () => Promise<any>
}